伤害计算全部采用Int,增加可破坏物件脚本-0

+++ Raindrop's Dream/Assets/Scripts/Action/HitAction.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public abstract class HitAction : MonoBehaviour
 {
-    public abstract void Hit(float _damage, HitEffect _hitEffect);
+    public abstract void Hit(int _damage, HitEffect _hitEffect);
 }
